+++ changelog.md
@@ -22,6 +22,7 @@ Codion Change Log
 - FilteredTableSortModel, add/removeListener methods replaced with methods that return the underlying event observer.
 - FilteredTableSearchModel, add/removeListener methods replaced with methods that return the underlying event observer.
 - FilteredComboBoxModel, add/removeListener methods replaced with methods that return the underlying event observer.
+- FilteredTableModel, add/removeListener methods replaced with methods that return the underlying event observer.
 ### is.codion.swing.common.ui
 - TemporalField, add/removeListener methods replaced with methods that return the underlying event observer.